Image_SR/Han/requirements.txt
gradio==4.44.1
huggingface_hub==0.23.5
pillow<11.0,>=8.0
super-image==0.1.7
torch>=1.9,<3.0
